+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Master
+* Use filename as a full metadata
 * Authorize only A-Za-z0-9._- for bucket name
 * Throw exception when some required defined properties are empty
 * Fix File getter return types